заменил массив на список в ProductBasket
