+++ src/org/skypro/skyshop/searchables/product/FixPriceProduct.java
@@ -1,4 +1,4 @@
-package org.skypro.skyshop.product;
+package org.skypro.skyshop.searchables.product;
 
 public class FixPriceProduct extends Product {
