+++ src/org/skypro/skyshop/searchables/product/SimpleProduct.java
@@ -1,4 +1,4 @@
-package org.skypro.skyshop.product;
+package org.skypro.skyshop.searchables.product;
 
 public class SimpleProduct extends Product {
     private int price;
